Make ShardsResolver separate kqp library (#9098)
